Extract todomvc-orbit-server into its own repo
