+++ app/data-strategies/event-logging.js
@@ -1,7 +1,7 @@
-import { EventLoggingStrategy } from "@orbit/coordinator";
+import { EventLoggingStrategy } from '@orbit/coordinator';
 
 export default {
   create() {
     return new EventLoggingStrategy();
-  }
+  },
 };
