+++ app/router.js
@@ -7,6 +7,6 @@ export default class Router extends EmberRouter {
 }
 
 Router.map(function () {
-	this.route('active');
-	this.route('completed');
+  this.route('active');
+  this.route('completed');
 });
