--- client/app/helpers/gt.js
@@ -1,7 +1,0 @@
-import { helper as buildHelper } from '@ember/component/helper';
-
-export function gt([n1, n2]/*, hash*/) {
-	return n1 > n2;
-}
-
-export default buildHelper(gt);
