--- client/app/helpers/pluralize.js
@@ -1,8 +1,0 @@
-import { helper as buildHelper } from '@ember/component/helper';
-import { pluralize } from 'ember-inflector';
-
-export function pluralizeHelper([singular, count]/*, hash*/) {
-	return count === 1 ? singular : pluralize(singular);
-}
-
-export default buildHelper(pluralizeHelper);
