--- client/package.json
@@ -1,74 +1,0 @@
-{
-  "name": "todomvc-ember-orbit-client",
-  "version": "0.0.0",
-  "private": true,
-  "description": "Example",
-  "repository": "",
-  "license": "MIT",
-  "author": "",
-  "directories": {
-    "doc": "doc",
-    "test": "tests"
-  },
-  "scripts": {
-    "build": "ember build --environment=production",
-    "lint": "npm-run-all --aggregate-output --continue-on-error --parallel 'lint:!(fix)'",
-    "lint:fix": "npm-run-all --aggregate-output --continue-on-error --parallel lint:*:fix",
-    "lint:hbs": "ember-template-lint .",
-    "lint:hbs:fix": "ember-template-lint . --fix",
-    "lint:js": "eslint . --cache",
-    "lint:js:fix": "eslint . --fix",
-    "start": "ember serve",
-    "test": "npm-run-all lint test:*",
-    "test:ember": "ember test"
-  },
-  "devDependencies": {
-    "@ember/optional-features": "^2.0.0",
-    "@ember/render-modifiers": "^1.0.1",
-    "@ember/test-helpers": "^2.2.0",
-    "@glimmer/component": "^1.0.3",
-    "@glimmer/tracking": "^1.0.3",
-    "babel-eslint": "^10.1.0",
-    "broccoli-asset-rev": "^3.0.0",
-    "ember-auto-import": "^1.10.1",
-    "ember-cli": "~3.25.0",
-    "ember-cli-app-version": "^4.0.0",
-    "ember-cli-babel": "^7.23.1",
-    "ember-cli-dependency-checker": "^3.2.0",
-    "ember-cli-htmlbars": "^5.3.2",
-    "ember-cli-inject-live-reload": "^2.0.2",
-    "ember-cli-sri": "^2.1.1",
-    "ember-cli-terser": "^4.0.1",
-    "ember-export-application-global": "^2.0.1",
-    "ember-fetch": "^8.0.4",
-    "ember-functional-modifiers": "^0.4.0",
-    "ember-load-initializers": "^2.1.2",
-    "ember-maybe-import-regenerator": "^0.1.6",
-    "ember-orbit": "^0.17.0-beta.5",
-    "ember-page-title": "^6.2.1",
-    "ember-qunit": "^5.1.2",
-    "ember-resolver": "^8.0.2",
-    "ember-source": "~3.25.1",
-    "ember-template-lint": "^2.18.1",
-    "eslint": "^7.20.0",
-    "eslint-config-prettier": "^7.2.0",
-    "eslint-plugin-ember": "^10.2.0",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "loader.js": "^4.7.0",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.2.1",
-    "qunit": "^2.14.0",
-    "qunit-dom": "^1.6.0"
-  },
-  "dependencies": {
-    "todomvc-app-css": "^2.2.0",
-    "todomvc-common": "^1.0.5"
-  },
-  "engines": {
-    "node": "10.* || >= 12"
-  },
-  "ember": {
-    "edition": "octane"
-  }
-}
