--- server/Procfile
@@ -1,1 +1,0 @@
-web: yarn start:heroku
