--- server/src/app.ts
@@ -1,3 +1,0 @@
-import App from './';
-
-module.exports = App;
