--- server/src/schema.json
@@ -1,10 +1,0 @@
-{
-  "models": {
-    "todo": {
-      "attributes": {
-        "title": { "type": "string" },
-        "completed": { "type": "boolean" }
-      }
-    }
-  }
-}
